refactor: separate collision to own file
